+++ app/src/main/java/com/contactsharing/beamit/resources/signup/SignupResponse.java
@@ -7,21 +7,21 @@ import com.google.gson.annotations.SerializedName;
  */
 public class SignupResponse {
     @SerializedName("user_id")
-    Long userId;
+    Integer userId;
 
     public SignupResponse(){
         //Required by the system
     }
 
-    public SignupResponse(Long userId) {
+    public SignupResponse(Integer userId) {
         this.userId = userId;
     }
 
-    public Long getUserId() {
+    public Integer getUserId() {
         return userId;
     }
 
-    public void setUserId(Long userId) {
+    public void setUserId(Integer userId) {
         this.userId = userId;
     }
